fix: 長考の時間・改行入りヘッダ・正規化エラーの型
- TimeFormat の分と秒を u16 にする。R-KIF-008 の「分」に上限は無く、
  持ち時間9時間の対局は 256分 を超える手を出す。u8 だと時間欄の
  パースが失敗し、opt がそれを None にして「時間の記載が無い手」に
  すり替わる。保存し直すと欄ごと消える。CSA の Duration も同じ理由で
  飽和にする（normalizer は既に飽和だった）
- ヘッダの値に改行があると、ヘッダの並びがそこで切れて以降が指し手でない
  行として読み飛ばされる。header は HashMap なので、同じ棋譜を保存する
  たびに消えるヘッダが変わる。R-KIF-004 のとおり1行に潰す
- ParseError::Normalize を String から NormalizeError にする。呼び出し側が
  「非合法手を含む棋譜（R-RULE-002 上は正常な入力）」と「壊れたファイル」を
  文字列一致でしか分けられなかった。consumer がそこで文字コードの
  取り違えを疑って別のデコードを試すのが、無言の全損の入口になっている
  （GAP-004 の誤字もここで消える）

+++ src/converter.rs
@@ -122,7 +122,7 @@ mod tests {
     #[test]
     fn a_node_without_a_move_line_writes_no_time_line() {
         use crate::jkf::{Color, Kind, MoveMoveFormat, PlaceFormat, Time, TimeFormat};
-        let ticks = |s: u8| Time {
+        let ticks = |s: u16| Time {
             now: TimeFormat { h: None, m: 0, s },
             total: TimeFormat { h: None, m: 0, s },
         };
